Create GRADES_FUNC
Create a function called grades that returns the grade from the parameter mark.
The parameter is a numerical value. returns: 'A' for mark 80+, 'B' for 70+, 'C' for 60+ and 'F' for less than 60
Create a loop to store 10 grades in a list based on a user's input of a mark and store the marks and students' names in two seperate list.
You should have three lists in total.
Output all the students' names, marks and grades.
Challenge: Can you prompt the user for a name and then output only that student's mark and grade?